Added TeacherControllerREST with GET/PUT/POST/DELETE HTTP methods Changed annotations in Teacher to send it as JSON Added Validation of Teacher Added NoTeacherWithSuchIdException and ExceptionHandler for it

+++ src/main/java/com/sbproject/schedule/services/implementations/TeacherServiceImpl.java
@@ -1,5 +1,7 @@
 package com.sbproject.schedule.services.implementations;
 
+import com.sbproject.schedule.exceptions.teacher.NoTeacherWithSuchIdException;
+import com.sbproject.schedule.models.Lesson;
 import com.sbproject.schedule.models.Subject;
 import com.sbproject.schedule.models.Teacher;
 import com.sbproject.schedule.repositories.TeacherRepository;
@@ -8,6 +10,7 @@ import com.sbproject.schedule.utils.Markers;
 import com.sbproject.schedule.utils.Utils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.aspectj.weaver.ast.Not;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -30,22 +33,43 @@ public class TeacherServiceImpl implements TeacherService {
 
     //TO DO - need to add name check probably
     @Override
-    public boolean addTeacher(String name, Set<Subject> subjects) {
+    public boolean addTeacher(String name, List<Subject> subjects) {
         teacherRepository.save(new Teacher(name, subjects));
         return true;
     }
 
-    //TO DO - need to add name check probably
     @Override
-    public void deleteTeacher(Long id) {
+    public Teacher addTeacher(Teacher teacher) {
+        teacher.setId(-1L);
+        return teacherRepository.save(teacher);
+    }
+
+    @Override
+    public boolean teacherExistsById(Long id) {
+        return teacherRepository.existsById(id);
+    }
+
+    @Override
+    public void deleteTeacher(Long id) throws NoTeacherWithSuchIdException{
+        if(!teacherExistsById(id)) throw new NoTeacherWithSuchIdException(id, "deleted");
         teacherRepository.deleteById(id);
-        logger.info(Markers.DELETE_TEACHER_MARKER,"Teacher has been successfully deleted!");
     }
 
     @Override
     public boolean updateTeacher(Long id, String name) {
         teacherRepository.save(new Teacher(id, name));
         return true;
+    }
+
+    @Override
+    public Teacher updateTeacher(Teacher teacher) throws NoTeacherWithSuchIdException {
+        if(!teacherExistsById(teacher.getId())) throw new NoTeacherWithSuchIdException(teacher.getId(), "updated");
+        return teacherRepository.save(teacher);
+    }
+
+    @Override
+    public Teacher getTeacherById(Long id) throws Exception{
+        return teacherRepository.findById(id).orElseThrow(() -> new Exception("Teacher with id '"+ id +"' not found!"));
     }
 
     @Override
